requirements.txt
alembic==1.4.3
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
netmiko==3.3.2
ntc-templates==1.6.0
paramiko==2.7.2
psycopg2==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyserial==3.5
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
scp==0.13.3
six==1.15.0
SQLAlchemy==1.3.20
tenacity==6.2.0
textfsm==1.1.0
Werkzeug==1.0.1
